test(Fixtures): rename Typescript fixtures and refctor to use factory functions
