+++ src/__fixtures__/article/simple.ts
@@ -1,0 +1,28 @@
+import { article, heading, paragraph } from '@stencila/schema'
+
+/**
+ * A very simple `Article` with only headings, paragraphs
+ */
+export default article({
+  title: 'The title',
+  content: [
+    heading({
+      depth: 1,
+      content: ['Heading one']
+    }),
+    heading({
+      depth: 2,
+      content: ['Heading two']
+    }),
+    paragraph({
+      content: ['Paragraph one.']
+    }),
+    heading({
+      depth: 3,
+      content: ['Heading three']
+    }),
+    paragraph({
+      content: ['Paragraph two.']
+    })
+  ]
+})
