+++ src/codecs/docx/docx.test.ts
@@ -1,4 +1,4 @@
-import articleSimple from '../../__fixtures__/article-simple'
+import articleSimple from '../../__fixtures__/article/simple'
 import { DocxCodec } from './'
 
 test('invertible', async () => {
