+++ src/codecs/json5/json5.test.ts
@@ -1,4 +1,4 @@
-import articleSimple from '../../__fixtures__/article-simple'
+import articleSimple from '../../__fixtures__/article/simple'
 import { Json5Codec } from './'
 
 test('invertible', async () => {
